DB corretto(rimosso cascade colloquio)

+++ Temp/src/main/java/com/example/demo/interfaces/ColloquioRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.demo.entities.Colloquio;
 import com.example.demo.entities.Intervistatore;
 
-public interface ColloquioRepository extends JpaRepository<Colloquio,Long> {
+public interface ColloquioRepository extends JpaRepository<Colloquio,Integer> {
 
 	public List<Colloquio> findByCandidatoNome(String nome);
 	public List<Colloquio> findByIntervistatoreNome(String nome);
